requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
supabase
PyJWT
python-dotenv